Added Background Image and experimenting with opacity
I can't get it to stop drawing after it's been clicked. This shouldn't
really ever matter because the screen will be switching on a click event
anyway, but it bothers me that I can't figure out why it's doing what
it's doing incorrectly. I feel this may be a symptom of an underlying
bug. That's why I created a whole new branch for this. I didn't want to
muck up the good version with my experiments.