Interaction between Notebook and Notepage
I think I have worked through all of the bugs that could come between
the interactions. Add the bugs to the bug list when you find them. The
only thing I did not to, which I am leaving to Allison, is saving and
restoring that information. Just realized, I forgot the enter button. I
will add that.